abstract dev server into separate file
